build: ➕ add @solidjs/router 📦
