+++ packages/habit/package.json
@@ -15,6 +15,7 @@
     "vite-plugin-solid": "^2.3.0"
   },
   "dependencies": {
+    "@solidjs/router": "0.5.1",
     "solid-js": "^1.5.1"
   }
 }
